Bad Type Declaration
Minor error fix before I sign off. I accidentally declared num for newfloat
as an int instead of a float.

+++ FinalCompilerProject/helper_functions.c
@@ -265,7 +265,7 @@ void pError(errorLevel el, char* s, ...)
  }
 
  struct ast *
- newfloat(int num)
+ newfloat(float num)
  {
 	 struct floatval *a = malloc(sizeof(struct floatval));
 	 if(!a) {
